Added resting experiment
This is a pre-cursor to the seamount test case. It is soooo
trivial it is hardly including. HOWEVER, a planned [small]
modificiation to laterally compensating T, S gradients will
provide a more meaningful test.

For now, the test is trivial - no motion should occur for flat
initial conditions and topography.