Changed PTR_ to ALLOCABLE_ in types.
The F90 standard did not allow allocatable arrays to exist in types, but the
F95 and later standards do.  We have found it to be more efficient to use
allocatable arrays in place of pointers wherever practical, and in this case
those pointers declared only in dynamic memory mode via the PTR_ macro can
always safely be treated as allocatable arrays.  No answers are changed.