Changed duratio of integrations (changes answers)
Reduced length of several test cases to speed-up the development
work flow. Previously, the 23 experiments took over 20 minutes in
total to complete (averaging 1 minute per test). Now the total is
around 8-9 minutes, with the longest test being just over 1 minute.

This DID change the timestats.* files but not the answers.