Changed behaviour of commented parameters (!PARAM)
If a log_param of a parameter preceeded with a bang (!), i.e.
logging of a a commented parameter, the behaviour used to be to
always write messages, even if the parameter was previously logged.

Now, we only log once, and if the second logging is inconsistent we
do NOT issue any messages. I've opted for this since reported parameters
(like SYMMETRIC_MEMORY_) can have multiple states and are not read in
at run-time but do need to be logged. -AJA