Changed format of floats so can be read back in
Some parameters, when written, had recurring digits in the formatted
log files. Namely,
DISSIPATION_N1 = 5.999999999999999E-04
INT_TIDE_DECAY_SCALE = 300.3003003003003

We now try writing at a lower number of significant digits and re-read
the ascii form of the number to see if it gives the same numeric value.
If not, we try more significant digits. This approach rounds 5.99999
to 6.0 which have the same numeric value.