Obsoleted NORMALIZED_SUM_OUT & streamlined code.
  Obsoleted the unused and archaic option NORMALIZED_SUM_OUT (what remains
reproduces what happened when this was true), and added code to properly
format all output of the times and number of timesteps over a larger range
of sizes, so that useful ascii output is now provided when n > 1000000, but
the output is the same as before for smaller values.  All test cases give
identical output to before.