Added the first version of the internal tide code.
  This is the initial check-in of MOM_internal_tides & MOM_internal_tide_input.
These add the code to treat the far-field propagation of the low-mode internal
tides and extracted the calculation of the energy source into the internal tides
out of MOM_set_diffusivity (it is now in both places).  Also added the run-time
option INTERNAL_TIDES to activate this code.  There may still be bugs in the
code that handles the refraction of the low-mode internal tides, and the output
has yet to be used to provide a mixing prescription or to modify the solution.
All answers are bitwise identical in the test cases.