Performance improvements to reproducing_sum.
  Decreased the cost of error handling on the input arrays in reproduce_sum.
Also added optional arguments to the 2-d version to either do a standard
non-reproducing sum or to do a reproducing sum with no checks on the range of
the input variables.  These would not be used in production, but are useful
for the unit-test driver that is being introduced to test and time this
capability.  All answers are bitwise identical.