Added config_src/unit_driver/MOM_sum_driver.F90
  This is the first unit driver for stand-alone tests of MOM6 capabilities.
In this case, the code carries out a number of global sums, using the MOM6
reproducing sum capabilty (and standard global sums for comparison).  Other
unit drivers are planned for the future, so this directory should not be
used in the list_paths arguments, to avoid trying to compile with multiple
mains; instead individual files should be listed.  The list of paths and files
required for this driver is:
  config_src/dynamic/ config_src/unit_drivers/MOM_sum_driver.F90 src/framework/ \
  src/core/MOM_grid.F90 src/initialization/MOM_grid_initialize.F90 shared/
This unit test driver does not affect the main MOM6 model.