Changed intent of u_cor_aux in MOM_continuity.F90
Both u_cor_aux and u_cor_aux had intent(out) but when the
caller passed the same array to another argument with inent(in)
the gnu compiler issued warnings. Changing intent(out) to intent(inout)
avoids the warnings.

The main purpose of the intent(out) was to document the intent rather
than restrict the compiler to preserve incoming values. intent(inout)
works but the functional intent is still intent(out).