Corrected a potential NaN in MOM_diag_to_Z.
  Changed the application of an Adcroft's rule division to avoid a later
division by 0 when the total ocean depth gets to be less than 0.5 m, by
recasting this fixed value to be a multiple of the thinnest layer depths,
which is consistent with the logic later in the code.  This only affects
diagnostics, but by avoiding (non-propagating) NaNs can permit certain types
of debugging.  The answers are bitwise identical in all test cases.