Corrected the Flather OBCs in continuity_PPM.
  Corrected the treatment of Flather OBCs in continuity_PPM when the input
and output thickness arrays are actually the same, by making a copy of the
input thickness.  Answers are bitwise identical in cases that worked, but
this is a step toward allowing the Flather OBCs to work even with the non-split
time stepping schemes.  Answers are bitwise identical in all test cases.