Started splitting the MOM_CS_type.
  Started splitting the single MOM_CS_type that is shared between MOM.F90 and
the various MOM_dynamics_....F90 files into two separate control structures,
for the top level routine and for the dynamic cores.  Answers are bitwise
identical, but this did entail a reordering of some of the parameter reads, so
the various MOM_parameter_doc files were rearranged.  For now, both the control
structures are used in the dynamics routines, but this will be changed with
later revisions.  These changes will not work with previous versions of the model.