Finished separating MOM_CS_type from MOM_dyn_CS.
  The control structure for the dynamic cores is now completely separate from
the overall control structure, opening the way for each dynamic core to have
its own control structure and to simplify the barotropic coupling while
keeping important legacy configurations (like CM2G).  Several interfaces and
a shared type are changed, so this is not backward compatible, but all
answers are bitwise identical with previous solutions.  The order in which
parameters are logged has changed, so the various MOM_parameter_doc files
have been updated.