Refactored MOM_tracer.F90 out of existence.
  Replaced the single module MOM_tracer with 3 new modules, MOM_tracer_registry,
MOM_tracer_advect, and MOM_tracer_hor_diff, and made corresponding changes to
all of the files that used MOM_tracer.  This changes the order in which certain
parameters are logged, so new versions of MOM_parameter_doc... were also
checked in.  The solutions are bitwise identical in all test cases.