Decalre version via #include "version_variable.h"
version = '$Id' and tagname = '$Name' have been revised to
work with git. $Id$ and $Name$ were specific to CVS.

o tagname has been deleted completely
o version is now set to _FILE_VERSION is the macro exists
o otherwise, versino =- 'unknown'

At some point in the future, the compile line will have the
macro set to the output of a script that returns the version
and status of each file.