Completed making diag_ptrs a private type.
  The last pointers to diagnostics were removed from the diag_ptrs type,
with the relevant diagnostic arrays made local to appropriate modules.  A new
interface, get_diag_time_end, allows the diag_ptrs type to be treated as
private to the MOM_diag_mediator module.  Several arrays are no longer available
to MOM_PointAccel, and have been removed.  There are now two diagnostics of the
vertical diffusivity: Kd is the diffusivity as used after boundary conditions
and CFL constraints are applied, and Kd_set is the diffusivity as it was
specified.  All answers are bitwise identical.