Replace grid with G for the grid variable.
  Some places used grid instead of G for the grid_type variable.  Almost every
instance where these are used are either as arguments or as grid%..., so it is
not hard to find all of these variables, even with a single-letter name.  This
is now consistent throughout the files under src/.  All answers are bitwise
identical.  Because this includes changes to the name of an element in a type,
this change is not compatible with previous versions of the code.