Added optional arg. dynamic to MOM_domains_init.
  The new dynamic argument to MOM_domains_init is used to force the domain not
to use static memory, as a step toward sharing code between MOM6 and SIS2.  Also
added a new optional argument, local_indexing, to get_domain_extent, to generate
the same indexing as is currently used in SIS when it is present and false.  The
answers are bitwise identical in all MOM6 test cases.