Added submodule pkg/CVmix
Let's see if anyone notices? :)

When cloning the MOM6 repository, pkg/CVmix will appear as an empty
directory.

*** IF YOU DO NOT WANT TO SEE CVmix, DO NOTHING ELSE ***

If you want to populate pkg/CVmix do:
  git submodule init
  git submodule update

Technical details:
The URL is currently
 ssh://cvs.princeton.rdhpcs.noaa.gov/home/aja/Repos/CVmix.git
which is a crontab-synced clone of the SVN repository at
 https://cvmix.googlecode.com/svn/trunk
The sync is hourly. We'll treat this as a read-only process (for now).
Ideally, we'll move this clone to a publicly accessible repository
(code.google or github) sometime soon.