Revised model of roundoff for non-negative sum
I worked out the actual bounds on numerical roundoff for a
non-negative summation and the upper bounds on the difference
between two such sums. I'm quite confident the analysis is
right because adjust the estimate downward a touch cause some
false positives.

This all applies to the regridding problem, checking that the
end points of two grids is consistent.

See
http://mom6dev.wordpress.com/2013/07/13/modeling-numerical-roundoff-for-regridding/
for details