Corrected array extents in MOM_diag_mediator.
  The symmetic memory array extents being passed to send_data were being
shifted in the wrong direction, but the right results were being obtained
because send_data only uses the sizes indicated by the extents and assumes
that halos are symmetric in memory.  These changes will also allow post_data
to accept arrays that are declared to be only on the computational domain.
All interfaces and solutions are unchanged.  -RWH