Answer change: switched site to fre-commands-bronx-5
CVS tag for the site directory updated to fre-commands-bronx-5.

The bronx-5 version now provides gnu.mk (previously not supplied by FRE)
and has the compiler option -fdefault-double-8 which was not in the GOLD/MOM6
setup. This option changes answers for models involving SIS.

Removing this option recovers the previous answers.