Bugfix in applyBoundaryForcing of diabatic()
netHeat was applied to the 3d temperature instead of a slice
array (T2d). The T2d then had SW added to it and the result
put into the 3d array. The result was that only SW was being
added to the model. Oops

Changed answers for global_ALE/z and single_column_z