Simplified & extended set_grid_metrics_from_mosaic
  Simplified the structure of the set_grid_metrics_from_mosaic code, including
properly filling all values in the halo regions, and avoiding the use of
complicated array-syntax constructs.  While some grid metrics at the edges
of the halo regions are now set properly (but were not before), this did not
change the answers in any existing MOM6 test cases.  The external interfaces
are unchanged.