Removed use of #override in single_column expts
The parameters F_0 and INPUTDIR were set in both MOM_input
and each of the MOM_override's provied in the location specific
sub-directories of INPUT (e.g. INPUT/BATS). The latter had to
have F_0 and INPUTDIR specified via an #Override derictive. This
meant that a further override was not possible which inhibited
Niki from running restart tests.

I removed F_0 and INPUTDIR from MOM_input and thus no longer need
the #override. Technically the MOM_override are no longer overrides
and should be renamed to something like MOM_locality and MOM_override2
renamed to MOM_override and left empty, but that might break something
else so I left this as is.

Fixes #1964