Added DIABATIC_FIRST option.
  Added a new run-time option to apply the diabatic processes and buoyancy
forcing before the dynamics.  This is enabled with DIABATIC_FIRST, but is off by
default.  Mass, heat, and salt are conserved with the new option as well as with
the old one, but it is not otherwise extensively tested.  (Also added lots of
spaces after commas and corrected some misspelled comments.)  Interfaces are
otherwise unchanged, and all solutions are bitwise identical.