Added PRANDTL_TURB and turbulent viscosity.
  Added code to use the turbulent diffusivity in Kd_turb as the basis for a
corresponding turbulent viscosity.  The default value of PRANDTL_TURB, 0., does
not change any answers, while setting it to a larger value appears to work
correctly in the DOME test case. In addition, vardesc is made publicly available
via MOM_restart, and Kd_turb and TKE_turb were registered to be a part of the
restart files.