Corrected bug in defineAxes for axesZL.
  The argument for defineAxes of the 1-d layer axis axesZL had been for the
interface axis, but is now corrected; this only affects very few diagnostics.
In addition, added code (from SIS2) so that the diagnostics work correctly even
when G%isd and G%jsd are not 1.  All answers are bitwise identical, with the
axis label and extent of a few diagnostics reverting to their proper values.