Rearranged ice_thm.F90 arithmetic for robustness.
  Rearranged several expressions in the ice thermodynamics code to (1) make them
robust to 0 input values (i.e., avoid unnecessary NaNs) and (2) avoid extra
divisions for efficiency.  This slightly changes the answers in all test cases
that use SIS2, and new references solutions are being checked in.  All external
interfaces are unchanged.  Answers are mathematically the same, but are not
bitwise identical.