Add BUFFER_LAYER_EXTRAP_LIMIT to bulk_mixed_layer.
  Added a new run-time flag, BUFFER_LAYER_EXTRAP_LIMIT, that both constrains the
detrainment into the interior when the temperature or salinity changes exceed
the values set by ALLOWED_DETRAIN_TEMP_CHG and ..._SALT_CHG, and limits the
density range over which extrapolation can occur when detraining into the
lightest interior layer.  Also added a check for negative layer thicknesses in
MOM_state_stats.  By default, this new limit is not exercised, but when
exercised it will change answers.