Corrected bug in MOM_regularize_layers.
  Corrected a bug in MOM_regularize_layers that in extreme circumstances would
lead to negative thicknesses when the mixed and buffer layers are within the
density range of the lightest interior layer.  This does change answers when
REGULARIZE_SURFACE_LAYERS is true, and new reference solutions have been checked
in for the coupled_AM2_SIS test cases.  This is a severe bug in code that is not
widely used, so we have left no way to reproduce the old, flawed answers.