*Distribute negative surface salt flux
A negative surface salt flux (associated with freezing sea-ice) was leading
to negative salinities. In the ALE code, there is no option to "adjust_salt"
by added salt and breaking conservation.

The salt flux was previously being extracted from only the top layer. Now
the salt is extracted from the top layer unless there is not enough salt,
in which case the excess flux is passed on to the next deepest layer. The
pass through continues to the bottom.

If any salt flux is unused it is simply lost and the model is then non-conserving.

Changes answers for MOM6z_SIS_025 only, though should change other experiments
on a long enough time scale.