More stable tridiagonal solver in ice_temp_SIS2.
  Recast the ice_temp_SIS2 tridiagonal solver to use only addition between the
terms for greater stability and accuracy, along the lines of what is used in
MOM6.  Additionally I recast the solution of the quadratic equation with the
SIS2 form of the thermodynamics for greater accuracy.  Finally, SIS1 had omitted
the solar heating of snow in its first temperature estimate, and this has now
been added back.  The differences in short test runs are small but detectable,
and there may be long-term climate differences due to the inclusion of the solar
heating of snow.  New reference solutions were checked in for the cases that use
SIS2 with SIS1_THERMODYNAMICS=False.