+*Recast ice_resize_SIS2 to use mass and enthalpy.
  Reformulated ice_resize_SIS2 to take mass and layer enthalpy as arguments in
place of thickness and temperature. This led to a significant simplification of
the code.  The answers are mathematically equivalent but not bitwise identical,
and the interfaces to ice_resize_SIS2 has been altered.  This change also
eliminated the unused functions emelt2temp and ecolumn_SIS2.  New reference
solutions were checked in.