Corrected SIS2 enthalpy conservation diagnostic.
  The enthalpy budget diagnostic in SIS2 had not been accounting properly for
the latent heat fluxes from the ocean when ice is entirely evaporated within a
time step.  This has now been fixed, and the seaice.stats files are now showing
heat conservation to ~17 decimal places in the coupled model runs, as well as
the ice-ocean runs.  In addtion, code was added to columns in which heat or
mass non-conservation was occurring when COLUMN_CHECK is true.  Also, a
negative value of ENTHALPY_UNITS are now interpreted to mean an inverse.  None
of these changes change answers, but they do change the coupled seaice.stats
files, and new reference seaice.stats files have been checked in.