Redefine heat content of water fluxes; code clean
This commit is the first to work through code changes associated
with reorganizing/redefining the heat content of liquid and frozen
precip and calving.
eliminate the diagnostic array tempxPmE
change many field names to be more succinct
add some new diagnostics
solo experiment conserves but coupled ocean_SIS (with bulkmixed)
fails to conserve.  So have a known bug to fix.