Water fluxes and corresponding heat fluxes
Water fluxes and corresponding heat fluxes now conserve for
the GLOBAL_SIS and global_ALE/z test cases.  The treatment
of water leaving the ocean is further documented and
hopefully more transparent to the analyst and developer.
There are some incomplete steps that need cleanup related to
heat content of rivers/calving.  We need to test this code
still.