*+Account for latent heat of trans-ice evaporation
  Corrected the heat from the ocean in SIS2 to properly account for the latent
heat flux required when evaporation occurs from the ocean in initially
ice-covered categories.  Also account for the difference in the heat required to
evaporate ice compared to water in this same instance in the sea ice heat
budget.  The ice budget again is closed to roundoff, but there are answer
changes in the coupled cases, and an interface change to accumulate_bottom_input
as the part_size is no longer used.  New reference solutions have been checked
in for the SIS2 coupled test cases.