*+Changed ice_temp_SIS2 to emphasize enthalpy.
  Extensively revised the final set of calculations in ice_temp_SIS2 to be based
on changes in the enthalpies, and not changes in the ice temperatures. These
solutions are mathematically identical, but will facilitate the use of more
sophisticated treatments of the ice thermodynamics later, such as allowing the
ice and water to have different heat capacities.  This includes adding three new
subroutines, lay_temp_enth, dTemp_dEnth, and e_to_melt_TS (the latter is
publicly visible), and removing the unneeded subroutine column_enthalpy.  This
changes answers (at the level of cascading roundoff) to all cases using SIS2
thermodyanmics, and new reference solutions have been checked in.