*+Replace SIS2 ice thicknesses with masses.
  Replaced several ice thickness variables (in m) with mass per unit area
variables (in kg m-2), including widespread replacement of IST%h_ice with
analogous expressions using IST%m_ice, and associated changes in the
multiplication or division by the density of snow or ice.  There are changes to
the meaning (and units) of several variables passed in via the interfaces, and
this changes answers (at the level of cascading roundoff) to all test cases
using SIS2; new reference solutions have been checked in.