Corrected heat content diagnostics for ALE mode
Corrected the heat content diagnostics for ALE mode.
As we are still using the "old" algorithm, we still
need to use the TempXpme diagnostic array.  All
heat_content arrays are filled diagnostically, but
they are not what the model uses prognostically.
When the algorithm is updated, TempXpme will be
removed, and the heat_content diagnostics will
also be used prognostically.