+Elevated set_visc_CSp to address issue #83
  Added a single copy of the set_visc_CS to the MOM_control_struct that is
passed into the initialization routines for each of the MOM_dynamics... modules.
In addition, the elements of MOM_dyn_split_RK2_CS are once again private, as
they always should have been.  This corrects the segmentation faults that were
arising when the GOLD_SIS was run with DIABATIC_FIRST = True (Issue #83), but
does not change answers in any cases that had previously run successfully.  There
are several interface changes associated with this update.