*New OBLdepth calculation based on POP approach. Ans changes*
Modifications to the OBLdepth calculation based now on what
is done by POP.  New results found to be insensitive to the
"correction" step, since the first pass now properly accounts
for the thickness of the surface layer.  This new approach
changes answers, but not by much.  New algorithm is
faster (since no need to perform the correction step) and
easier to understand.