+*Use bt_cont_type to limit corrective mass fluxes.
  Added code to use information from the bt_cont_type to stably allow for larger
corrective mass sources to the barotropic solver.  These corrective mass sources
are one way that precipitation and evaporation can be realized in the barotropic
solver (when the thermodynamics is first), and they keep the barotropic solver's
free surface height consistent with the sum of the layer thicknesses.  These had
previously been limited based on the initial depth of the ocean (G%bathyT), but
now are based on a CFL limit on the transport that can be realized by the sum of
the layers or the mass within the cell.  As these limits were not being
encountered in the test cases, the reference solutions do not change, but
answers can change in longer runs with USE_BT_CONT_TYPE true.  A new run-time
argument, BT_CONT_CORR_BOUNDS, has been introduced to allow the old solutions
to be recovered.