New option FIXED_OBLDEPTH; change default KPP_IS_ADDITIVE to True
Added a new option FIXED_OBLDEPTH with parameter FIXED_OBLDEPTH_VALUE.
This option is of use to help test across models using CVMix/KPP.
The option over-rides the results from CVMix/KPP OBL depth, and
replaces that depth by the constant depth FIXED_OBLDEPTH_VALUE.

Changed the default for the option KPP_IS_ADDITIVE from False
to now be True.  The True value is consistent with what we do
in OM4, and it aligns with GFDL use of the KPP scheme.  We should
thus be setting the default to align with such practices.