**New option for depth scale in Rib calculation
New option for depth scale used in Rib calculation.
Default RIB_DEPTH_RATIO=0.95 means that we are
computing a bulk Ri 0.95 smaller than with the
earlier code with ratio=1. This new default changes
answers in a way that is desired (slighly deeper
OBL depth).  The change corresponds to the POP
approach.  It is justified from a finite volume
interpretation of the depth of the surface layer.