Modified MOM to enable long thermodynamic steps
  Made a number of non-answer-changing modifications that move MOM toward allowing
thermodynamic timesteps to be longer than coupling timesteps:
 - Added a counter of the number of times the heat, mass, and salt fluxes in the
forcing type have been augmented since they were last used; zero out this
counter after the fluxes are used to drive diabatic forcing; and use a negative
value of this counter to indicate that a forcing type needs to be have its arrays allocated.
 - Added the run-time argument THERMO_SPANS_COUPLING to enable the thermodynamic
timestep to span multiple coupling timesteps.
 - Added a variable, CS%dt_trans, to the MOM_control_struct that records the accumulated time since the last tracer transport.
 - Expanded the range of forcing fields that allocate_forcing_type is able to
allocate with 2 new optional arguments, and use this to shorten the duplicated
code that does the forcing type allocation.
 - Added commented out code at the top-level that will later enable the
averaging of fluxes across multiple timesteps.
 - Initialize area_surf to -1 in the surface_forcing_CS, and use this to test
whether the area needs to be initialized, which allows first_call to be
eliminated.