Added H_subroudoff to permit Angstrom=0.
  Added G%H_subroundoff at various places in the numerators and denominators of
expressions in vert_fill_TS and mixedlayer_detrain_2, and changed
adjustEtaToFitBathymetry to avoid trying to dilate massless water columns.  All
of these changes were causing errors when Angstrom=0, but when Angstrom is set
to a small positive values (e.g., 10^-10 m), as in the existing test cases, the
answers are bitwise identical.