+Revised interface to absorbRemainingSW
  Changed several of the arguments to absorbRemainingSW into optional arguments,
reflecting how they are actually used.  In the simplest, Z-coordinate, version
of calls to this routine, it allows several unused variables not to be declared
or defined, and it also conceptually simplifies calls to this routine in this
case.  The interface to sumSWoverBands has been similarly simplified, although
in this case it was only called once, so these unused arguments are simply
eliminated.  All answers are bitwise identical.