+Added MOM_diapyc_energy_rec.F90
  Added a new module, MOM_diapyc_energy_rec.F90, that does an implicit
calculation of the potential energy change due to diapycnal diffusion.  This is
tested with a testing driver, and the answers calculated 4 different ways
(upward or downward and after a tridiagonal solver or as a part of the solver)
agree to within roundoff.  This code is not yet called as a diagnostic, but
it is envisioned that it will be, and that this code is the basis for the new
implicit energetically-based diffusive boundary layer scheme that is being
developed.