Added find_PE_chg to MOM_diapyc_energy_req.F90.
  Added a new subroutine, find_PE_chg, that calculates the changes in potential
energy due to mixing at an interface for either upward or downward passes, along
with calls that can be used to verify that it is correct.  This new subroutine
will be reused for the new energy-based boundary layer mixing approaches. This
code is not called by the main branch of the code, so all solutions are bitwise
identical.