+Moved wave_speed_init into MOM.F90.
  Relocated the initialization call for MOM_wave_speed up into the main MOM
initialization block, and added the wave_speed_CS as an argument to VarMix_init
and MOM_diagnostics_init. This change ensures that the wave_speed parameters are
only logged once and in the same order regardless of which diagnostics or
parameter settings are used, and that all MOM_parameter_doc files will be the
same regardless of the diag_table that is used.  In addition, the unused run-
time argumentes CONV_DECAY and HMIX_MIN are no longer read or logged by
MOM_energetic_PBL.  This changes some interfaces, but all solutions are bitwise
identical.