*Restricted use of CORRECT_ABSORPTION_DEPTH
  Eliminated the use of AdjustAbsorptionDepth with ALE models, and in models
with a bulk mixed layer change the way that CORRECT_ABSORPTION_DEPTH is applied
so that the water above is never heated at a faster rate than the water in the
layer that is being heated.  This does change answers slightly in many test
cases with penetrating shortwave radiation, and new reference solutions have
been checked in.