*Added options to improve barotropic transport
  Added the new run-time options ADJUST_BT_CONT and BT_USE_VISC_REM_U_UH0 that
will improve the consistency between the barotropic transport and the transport
that is realized by the layer equations.  This includes adding the new
subroutine adjust_local_BT_cont_types.  Also added a new run-time argument,
CONT_PPM_MARGINAL_FACE_AREAS, that will allow the selection of the marginal or
transport averaged face areas when setting the layer weights for the barotropic
solver via the BT_cont_type. In addition, added a number of diagnostics that can
be used to evaluate the fields in the BT_cont_type, primarily for debugging
purposes.  By default, none of these code changes change answers, but new
versions of the MOM_parameter_doc.all and available_diags files were checked in.