*Account for negative TKE in energetic_PBL.
  It is possible to get negative TKE in energetic_PBL due to the energetic
demands of mixing the surface buoyancy forcing over the topmost layer or due to
shortwave radiation.  In these cases, the cube root  of a negative number was
returning NaNs with some compilers, leading to model failures.  The new code
sets the TKE to 0 in the topmost internal interface and makes sure that vstar
and diffusivities derived from vstar are properly set to 0 in this situation.
This can change answers, especially if MSTAR is relatively small, but it did
not change answers in any of our test cases.