+Added module MOM_hor_index and hor_index_type.
  Added a new module with a type to describe the domain-decomposed horizontal
index ranges, added a new sub-element of this type to the ocean grid type, and
used this type to replace the ocean_block_type variables in the grid.  In many
places, the ocean block_type variables "B" were replaced with the more easily
grepped-for horizonal index type "HI". All answers are bitwise identical, but
there are new interfaces.