*Revamped filtered_grid_motion with varying rates
  Completely revised the algorithm in filtered_grid_motion when there is a
non-uniform damping rate so that when there is an ordered list of initial and
target interface depths, the resultant list is also ordered.  The answers are
the same for a uniform damping rate.  Also corrected the order of arguments to
two calls to filtered_grid_motions, so that they make sense and will work for
non-uniform damping rates.  Although this changes answers in configuations that
did not previously work, all of the answers in the existing test cases are
bitwise identical.