*Changed default for OCEAN_SURFACE_STAGGER to "C"
  Changed the default setting for the staggering of the ocean surface velocities
that are reported to the coupler and sea-ice from "B" to "C", reflecting the
native discretization of MOM6.  This would change the solutions for test cases
that use the default values, so new lines were added to many MOM_input files to
avoid this in the test cases.  By default answers do change, but by design they
have not changed in any of the existing test cases.