Added fluxes%p_surf_SSH
  Added a new array for the surface pressure as used in the sea surface height
corrections, to distinguish it from the full surface pressure, which should be
used in thermodynamics and other calculations.  Also corrected a grammatical
error in the comments describing USE_LIMITED_PATM_SSH, slightly changing a few
MOM_parameter_doc files.  Two (commented out) bug corrections are deferred in
the call to make_frazil and in the code calculating rigidity_ice_[uv] so that
the existing solutions are bitwise identical.