+Moved get_flux_units to MOM_verticalGrid
  Moved the primary initialization of 11 variables related to the vertical grid
from MOM_grid.F90 to MOM_verticalGrid.F90. Also moved the routines to determine
Boussinesq or non-Boussinesq units of thicknesses and fluxes into the vertical
grid module, and changed an argument from a grid_type variable to a
verticalGrid_type variable.  This changes the order in which several parameters
are logged in the MOM_parameter_doc files, but all answers are bitwise identical.