Changed the default for DEPTH_LIST_MIN_INC
  Changed the default for DEPTH_LIST_MIN_INC from ANGSTROM_Z to 1.0e-10.  These
are the same values in many cases.  This parameter affects the diagnosis of
available potential energy, and hence can change answers in the ocean.stats
files, but it does not actually change the model's solutions.  All answers are
actually bitwise identical, and are reported as such in the test cases.  New
MOM_parameter_doc.all files have been checked in.