+Revised ocean_grid_type to harmonize with SIS2
  Reoredered the elements of the ocean_grid_type, removed the axis_units
element, and added sin_rot and cos_rot all for consistency with the code that is
to be shared with SIS2.  As part of this, a duplicate read of GRID_CONFIG was
elminated, but this causes the units of DISK_RADIUS as documented in the
circle_obcs MOM_parameter_doc.all to change from "k" to "kilometers".  All
answers are bitwise identical.