Repaced NIMEM_ with SZI_ in user code
  The SZI_(G), etc., macros are the ones that are supposed to be used to specify
array sizes for arguments, but NIMEM_, etc., were being used instead, which
would cause errors if the model were ever run without G%isd=1.  The proper
argument array size macros are now used in the code in the user directory.  This
change also required that G and GV be defined as arguments before they are used,
so the order of a few lines had to be changed in many places.  All answers are
bitwise identical.